Don't delete result message if there were errors

+++ src/plugins/commander/commands/clear.js
@@ -133,8 +133,10 @@ class ClearCommand extends ModCommand {
                     confirmation.delete(),
                 ]);
 
-                await this.wait(5000);
-                await result.delete();
+                if (!failures) {
+                    await this.wait(5000);
+                    await result.delete();
+                }
 
                 break;
             case this.CROSS:
